// Winnocare-App/Winnocare/src/app/model/user.ts
export class User {
    userName: string;
    password: string;
    firstName: string;
    lastName: string;
    age: string;
    gender: string;
    country: string;
    phoneNumber: string;
    email: string;
    emergencyContact1: string;
    emergencyContact2: string;
    doctorContact1: string;
    doctorContact2: string;
}